WL-561 Restoring a resource never invokes an e-mail notification now.
In the scenario where a maintainer had accidentally deleted a resource and
then restored it, e-mailing all site members of this fact and thus alerting
them to their mistake is clearly highly undesirable! E-mail notification
was actually set to optional, but unless a user has their preferences set
to 'Do not send me low priority resource notifications' (not the default!)
they would get an e-mail. It does not really seem conceivable that anyone
would want this anyway however ;-).