BSGAPP-50: Resources should indent with a class
Fix the indentation starting at one.

The reason for the file depth to start from 1 was a limitation by the velocity template language. There is no support of for loops, but there is a support for "foreach" loops.
The code was #foreach($temp in [0..${item.depth}]) but [0..${item.depth}] create an array with every number from 0 to ${item.depth} (there is only one), this is why it had at least one level of indentation

To fix that, [1..${item.depth}] has been used, unfortunately for velocity [1..0] means every number from 1 to 0 (so it's 1 and 0). So the final code is now : #if(${item.depth}!=0)#foreach($temp in [1..${item.depth}])

This takes care of the case where 0 is used, and counts from level 1.

It's not really nice, but it gives a more accurate representation in resources.