WL-3190 Rename existing getLongAccessLabel
This is probably a bad move because I am creating another public method
with the exact same type signuture. That said it is used once in the UI.

The benefit is that we now have a much more descriptive method name and
I've thrown in some documentation because I'm nice like that.