WL-3132 Show Missing Group(s) in list
Currently shows Select Group(s), but if there are no groups available then
this indicates that a group must be missing.

This is quite naive because there might be one group remaning besides the
deleted one but this is the easiest to implement here. Will provide more
information in the Edit Properties page.