WL-3132 Edit properties has a missing warning
Now if there is a disagreement with the number of groups that are defined
on the entity and the number of groups on the listitem it will show a
message that indicates as such.

The message appears like a failed validation on a field with a red message
and an exclamation mark.