WL-3132 Show missing message even without table
Currently the message that groups are missing will only be displayed if the
table to add groups is going to be shown. Presumably so that the code is
all logically grouped together.

It is possible that no groups are missing and no groups are available, in
which case the table and the message is not displayed. This logic needs to
be separated out.