BSGAPP-1 - WP2.2 - Add new meta data fields BSGAPP-2 - WP2.1 - Store meta data in sites BSGAPP-3 - WP2.4 - Assign meta data to resources
WL-1955 PermissionException with metadata.json

The advisor checks content rights based on a reference, not an Id

WL-38 : Use properties to get Dublin Core group name

WL-2330 Support boolean metadata type.

This adds support for a boolean metadata type which can be used by Blavatnik for marking content as essential reading. I also refactored the JSON parser out into a helper so that it's easier to test the parsing of a config file.